separate product-name-cell component, and add placeholder for empty thumnail
