+++ src/app/catalogs/[catalog_id]/components/product-list/index.ts
@@ -1,1 +1,2 @@
 export { ProductList } from './product-list'
+export { ProductNameCell } from './product-name-cell'
